introduce MemberSettingTemplate
