+++ packages/atlas/src/views/viewer/MembershipSettingsView/MembershipPublicProfile/MembershipPublicProfile.styles.ts
@@ -1,12 +1,7 @@
 import styled from '@emotion/styled'
 
 import { ActionBar } from '@/components/ActionBar'
-import { LayoutGrid } from '@/components/LayoutGrid'
 import { sizes, zIndex } from '@/styles'
-
-export const StyledLayoutGrid = styled(LayoutGrid)`
-  row-gap: ${sizes(8)};
-`
 
 export const Wrapper = styled.div<{ actionBarHeight: number }>`
   padding-bottom: ${({ actionBarHeight = 0 }) => actionBarHeight}px;
