+++ packages/atlas/src/views/viewer/MembershipSettingsView/MembershipSettingTemplate/index.ts
@@ -1,0 +1,1 @@
+export * from './MembershipSettingTemplate'
